Refactor page props declaration
