+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Homepage from "src/pages/Homepage/Homepage";
+import Homepage from "src/pages/Homepage";
 
 export default function HomePage() {
   return <Homepage />;
