--- src/pages/Homepage/type.ts
@@ -1,9 +1,0 @@
-import GameVersion from "../../types/gameVersion.type";
-import { GameData } from "../../types/pokemonData.type";
-
-type HomepageProps = {
-  gameData: GameData;
-  chooseGameVersion: (version: GameVersion) => void;
-};
-
-export default HomepageProps;
