--- src/pages/Leaderboard/type.ts
@@ -1,8 +1,0 @@
-import { LeaderboardTotal } from "../../types/leaderboardData.type";
-
-type LeaderboardProps = {
-  leaderboardData: LeaderboardTotal;
-  difficulty: string;
-};
-
-export default LeaderboardProps;
